requirements.txt
Flask==3.0.0
cohere==5.15.0
python-dotenv==1.0.1
gunicorn==22.0.0
joblib==1.4.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.4.2
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
